move api signing logic into apiSignging
